+++ Rakefile
@@ -44,20 +44,11 @@ def version_greater_than(tag1, tag2)
 end
 
 namespace :test do
-  Rake::TestTask.new(:integration) do |t|
-    t.libs << 'test'
-    t.libs << 'lib'
-    t.test_files = FileList['test/integration/{**/,}*_test.rb']
-  end
   Rake::TestTask.new(:unit) do |t|
     t.libs << 'test'
     t.libs << 'lib'
-    t.test_files = FileList.new('test/**/*_test.rb') do |fl|
-      fl.exclude(/integration/)
-    end
+    t.test_files = FileList.new('test/**/*_test.rb')
   end
-  desc 'Run all tests'
-  task :all => [:unit, :integration]
 end
 
 namespace :docs do
